+++ src/main/java/com/example/gerenciamentoDeContas/service/UsuarioService.java
@@ -1,6 +1,8 @@
 package com.example.gerenciamentoDeContas.service;
 
 import com.example.gerenciamentoDeContas.exception.EntityNotFoundException;
+import com.example.gerenciamentoDeContas.model.Dto.UsuarioRequest;
+import com.example.gerenciamentoDeContas.model.Dto.UsuarioResponse;
 import com.example.gerenciamentoDeContas.model.UsuarioModel;
 import com.example.gerenciamentoDeContas.repository.IUsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,21 +11,34 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Service
 //Lógica e validações
 public class UsuarioService {
-
-
     @Autowired
     private IUsuarioRepository iUsuarioRepository;
 
-    public UsuarioModel cadastrarNovoUsuario(UsuarioModel usuarioModel) {
-        return iUsuarioRepository.save(usuarioModel);
+    public UsuarioResponse cadastrarNovoUsuario(UsuarioRequest usuarioRequest) {
+        UsuarioModel usuarioModel = new UsuarioModel();
+        usuarioModel.setNomeUsuario(usuarioRequest.getNomeUsuario());
+        usuarioModel.setDataDeNascimento(usuarioRequest.getDataDeNascimento());
+        usuarioModel.setEmail(usuarioRequest.getEmail());
+        usuarioModel.setCpf(usuarioRequest.getCpf());
+        iUsuarioRepository.save(usuarioModel);
+
+        return new UsuarioResponse(usuarioRequest.getNomeUsuario(),
+                usuarioRequest.getDataDeNascimento(),
+                usuarioRequest.getEmail(),
+                usuarioRequest.getDataDeNascimento());
+
     }
 
-    public List<UsuarioModel> exibirUsuarioCadastrado() {
-        return iUsuarioRepository.findAll();
+    public List<UsuarioResponse> exibirUsuarioCadastrado() {
+        List<UsuarioModel> usuarioModelList = iUsuarioRepository.findAll();
+        return usuarioModelList.stream().map(obj -> new UsuarioResponse(obj.getCodigo()
+                ,obj.getNomeUsuario(), obj.getDataDeNascimento(), obj.getEmail())).collect(Collectors.toList());
+
     }
 
 
@@ -38,6 +53,9 @@ public class UsuarioService {
     }
 
     public void deletarUsuario(UUID codigo) {
+        if (!iUsuarioRepository.existsById(codigo)) {
+            throw new RuntimeException("Erro, id não encontrado");
+        }
         iUsuarioRepository.deleteById(codigo);
     }
 
@@ -45,3 +63,4 @@ public class UsuarioService {
         return iUsuarioRepository.existsById(codigo);
     }
 }
+
